Backend/Connection/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
fastapi
uvicorn
mysql-connector-python
python-dotenv
pyjwt
requests
